+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Changed
 - templates: fixed minor bug in Java output
+- templates: fixed minor bug in output of companion object
 
 ## [0.94] - 2023-09-17
 ### Changed